Improved warnings and errors for multiple UDF definitions (#2961)
There is logic to detect if a UDF is overriding an existing function.
However, it doesn't properly differentiate between overriding a builtin
function, which should be allowed, from redefining a UDF with the same
name, which should not be allowed. Worse, replacing the UDF appears to
work but the old definition is retained.

Today's behavior:
```
>> f():Number = 3;

>> f():Number = 4;

>> f()
3

>> Abs(x:Number):Number = -x;

>> Abs(9)
-9

>> Abs(x:Number):Number = -x * 2;

>> Abs(9)
-9
```