Support := syntax for named formulas, with compatibility option for no colon (#2966)
We decided last year that `:=` should be used instead of `=` for
defining UDTs. We did this because we wanted to rationalize the syntax
of Named Formulas:
```
x = 3; // no colon, and no way to express return type
```
with UDFs:
```
x(): Text = 3; // colon, and return type is required
```
To rationalize, we should make the syntax consistently `:=`, with or
without a type specification.

Since Canvas already has `=` syntax in GA and has been in use for years,
we have not changed it there. But we have not GA'd the REPL or other
products that may support UDFs, and we should fix this before then.

This PR introduces a new ParserOption to support the `=` syntax in
Canvas, _in addition_ to the new `:=` syntax. This option is off by
default for other Power Fx hosts. But we would turn this option on for
Canvas.