Add Simple Expression constraint (#2980)
Some properties that depend on ThisItem scope will have significant
performance costs to async operations. We've discussed and landed on
restricting the set of things that can be accessed from those properties
to just sync functions, the thisitem scope, enums, and operators.

This PR is a first pass at this, but it won't be immediately exposed to
customers, so we'll iterate on it.